Added signing out functionality, all tests pass
