+++ app/views/layout.erb
@@ -7,6 +7,10 @@
 
     <% if current_user %>
         Welcome, <%= current_user.email %>
+        <form method="post" action="/sessions">
+            <input type="hidden" name="_method" value="delete">
+            <input type="submit" value="Sign out">
+        </form>
     <% end %>
 
     <% if flash[:notice] %>
