+++ spec/user_management_spec.rb
@@ -20,17 +20,6 @@ feature "User signs up" do
     expect(page).to have_content("This email is already taken")
   end
 
-  def sign_up(email = "test@test.com", username = "thrillhouse", name = "sean", password = "pray4mojo", password_confirmation = "pray4mojo")
-    visit '/users/new'
-    expect(page.status_code).to eq(200)
-    fill_in :name, :with => name
-    fill_in :username, :with => username
-    fill_in :email, :with => email
-    fill_in :password, :with => password
-    fill_in :password_confirmation, :with => password_confirmation
-    click_button "Sign up"
-  end
-
 end
 
 feature "User signs in" do
@@ -52,11 +41,18 @@ feature "User signs in" do
     sign_in('test@test.com', 'wrong')
     expect(page).not_to have_content("Welcome test@test.com")
   end
+end
 
-  def sign_in(email, password)
-    visit '/sessions/new'
-    fill_in 'email', :with => email
-    fill_in 'password', :with => password
-    click_button 'Sign in'
+feature "User signs out" do 
+
+  before(:each) do 
+    User.create(:email => "test@test.com", :password => "test", :password_confirmation => "test")
+  end
+
+  scenario "while being signed in" do
+    sign_in("test@test.com", "test")
+    click_button "Sign out"
+    expect(page).to have_content("Good bye!")
+    expect(page).not_to have_content("Welcome, test@test.com")
   end
 end
